Allow the user to checkout or return a book to the librarian
I added more commands that will let the user interact with the librarian
more. In addition, I added more messaging to help keep the user informed
of what is happening. I noticed one of my tests was failing and wasn't
properly using the should_panic decoration, so I fixed that.